backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
pgvector==0.3.6
bcrypt==4.2.0
python-dotenv==1.0.1
Werkzeug==3.0.4
gunicorn==23.0.0
pdfplumber==0.11.4
PyPDF2==3.0.1
openai==1.55.3
anthropic==0.38.0
google-generativeai==0.8.5
numpy==1.26.4
tiktoken==0.7.0
aiohttp==3.10.10
